+++ s/browser/host.ts
@@ -3,10 +3,10 @@ import {Sub} from "@e280/stz"
 
 import {SignallerApi} from "../signaller/api.js"
 import {ConnectOptions, StdCable} from "./types.js"
-import {connect, SparrowConnect} from "./connect.js"
+import {connect, Connect} from "./connect.js"
 import {AgentInfo, SignallerStats} from "../signaller/types.js"
 
-export class SparrowHost extends SparrowConnect {
+export class Host extends Connect {
 	constructor(
 			signaller: SignallerApi["v1"],
 			self: AgentInfo,
@@ -19,9 +19,12 @@ export class SparrowHost extends SparrowConnect {
 	}
 }
 
+/** @deprecated renamed to Host */
+export class SparrowHost extends Connect {}
+
 export async function host<Cable = StdCable>(options: ConnectOptions<Cable>) {
 	const {signaller, self, stats, onStats, close} = await connect(options)
 	const invite = await signaller.createInvite()
-	return new SparrowHost(signaller, self, stats, onStats, close, invite)
+	return new Host(signaller, self, stats, onStats, close, invite)
 }