+++ s/index.ts
@@ -1,21 +1,8 @@
 
-export * from "./browser/std/cable.js"
-export * from "./browser/std/options.js"
-export * from "./browser/std/rtc-configurator.js"
-export * from "./browser/std/url.js"
-
-export * from "./browser/utils/conduit.js"
-export * from "./browser/utils/data-channeller.js"
-export * from "./browser/utils/report-connectivity.js"
-
-export * from "./browser/connect.js"
-export * from "./browser/host.js"
-export * from "./browser/join.js"
-export * as Sparrow from "./browser/sparrow.js"
-export * as default from "./browser/sparrow.js"
-export * from "./browser/types.js"
+export * from "./browser/index.js"
+export * as Sparrow from "./browser/index.js"
+export * as default from "./browser/index.js"
 
 export * from "./signaller/types.js"
-
 export * from "./tools/id.js"