+++ CHANGELOG.md
@@ -8,6 +8,13 @@
 
 ## v0.2
 
+### v0.2.13
+- 🍏 fix the build lol
+- 🔶 rework exports, notably doing some renames with deprecated alias for old names
+  - `SparrowHost` => (`Host`, `Sparrow.Host`)
+  - `SparrowJoin` => (`Join`, `Sparrow.Join`)
+  - `SparrowConnect` => (`Connect`, `Sparrow.Connect`)
+
 ### v0.2.12
 - 🍏 update dependencies